python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv==1.0.0
openai>=1.0.0

# UIED dependencies (minimal - OCR removed)
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
tqdm>=4.65.0

# Optional: Advanced classification
# keras>=2.15.0
# tensorflow>=2.15.0

